+++ libs/@guardian/source-development-kitchen/.storybook/manager.ts
@@ -1,0 +1,6 @@
+import { addons } from '@storybook/manager-api';
+import { createTheme } from '../../../../configs/storybook/theme';
+
+addons.setConfig({
+	theme: createTheme('@guardian/source-development-kitchen'),
+});